Add files via upload
This GitHub repository contains a small banking system implemented in C++. Please note that this project is not a fully functional or optimized code but rather created for educational purposes and fun. If you are looking for an optimized version of the code, please let me know, and I can provide it separately.

The banking system implemented in this project showcases various features and functionalities typically found in a basic banking system. It includes the ability to create customer accounts, perform transactions such as deposits and withdrawals, check account balances, and display transaction history.

While the primary focus of this project is to demonstrate fundamental programming concepts and techniques in C++, it can serve as a starting point for beginners to understand the basics of object-oriented programming, file handling, and so on.